Code for checking dot product of finger directions
This will eventually be used to determine if the fingers are moving in the same direction (for dragging), in the opposite direction (for pinching) or perpendicular to their relative directions (for rotation). But it needs more work due to a quirk in how the positions are updated that prevents the dot product from registering properly.