Simplified code for determining old vs. new values in TwoFingerTouchControls
To reduce events being unnecessarily invoked, I created code that determined if the drag, pinch or rotate values had changed. The previous method was very roundabout and involved extra variable references and a more complicated system for updating which buttons are pressed. Since the original purpose of that more complex system was made redundant, I opted to refactor this functionality to be much simpler and easier to read.