Improved differentiation between pinching and rotation to be more forgiving.
The 'pinching' bool is determined by checking if the fingers are moving in opposite directions, but the perpendicular bools are both false, meaning they're directly moving towards or away from each other. The 'rotating' bool would instead check that the perpendicular bools were both true. I found that this caused more jittery input since only one might return true, even if the player is clearly doing a rotation.
To fix this, I kept the 'pinching' detection requirements equally strict, since it's a simpler movement to check for, but made it so the 'rotating' detection requirements will only need to check if one of the perpendicular bools is true. This way if the player makes a rotation move that isn't perfect but obviously isn't a drag or pinch, the code will still detect a rotation without mixing anything up.