Pan controls in ObjectView, using TouchDrag
This function uses ScreenToWorldPoint with exact screen positions to determine the exact world distance the object needs to move. This produces a movement that is almost 1:1 with the player's finger movement, although it does require ignoring the Vector2 parameter of TouchDrag's event and substituting a more customised listener.

The code doesn't change the rotation, so the viewed object will appear to 'rotate' as its position shifts relative to the camera's field of view, but I haven't decided yet if this is a problem that needs to be fixed. Overall, I am pleased with this code.