MultiTouch class for easily obtaining and tracking multiple touch values
This class allows multiple touch inputs to be obtained using the new input system, without the hassle of setting up an InputActionAsset or a PlayerInput.
This script will set up multiple InputActions to register touch inputs – one for registering contact and one for the finger position, multiplied by the supported number of touches. By automating it, I can easily set up a large number of them to support as many separate touches as the device will allow (or 10, as allowed by the input system), which would be much more obnoxious to set up manually due to the time expended, the separate functions having to be set up for each finger, and the possibility of making errors and not assigning variables and interactions properly.
The script also seamlessly tracks the number of active touches and displays their positions in an array, without needing to worry about which specific ones are being registered in the array, which makes it much easier to execute two-finger commands without worrying about their current fingers not being counted because they aren’t technically the first and second fingers. If specific touch inputs still need to be obtained, the InputActions can still be accessed through a public property.
I previously experimented with a version of this that used EventSystems interfaces – IPointerDownHandler, IPointerUpHandler, and IDragHandler. However, due to a quirk in the way touches are added and removed during that system, it was possible for the player to press or release multiple fingers in a single frame, and have only one change be registered. This resulted in phantom touches that the code thought were still being pressed. This new system swiftly avoids that problem because each touch is being tracked separately.
Overall, I’m quite pleased with it!