Bugfix - delta resets in OneFingerDragInput and TwoFingerTouchInput would not work on first use
After a lot of aggravating debugging, I realised this was due to the way the contact interactions were registered. When the player presses a finger down, it runs the contact action's event and then the position action's event immediately afterwards. Due to running twice, once before the new position value is actually registered, the delta reset check was unable to catch the new value and assign it to the 'old' variables used to determine delta values.

I fixed this by changing the contact action interactions to only register when being released. This won't prevent any inaccuracy when registering new presses, since when OnValuesChanged() runs from the position action event immediately afterwards it still updates the data appropriately.